move messageFooter out of reusable folder
